requirements.txt
certifi==2021.10.8
paho-mqtt==1.6.1